src/requirements.txt
Django==2.2.10
django-model-utils==4.0.0
djangorestframework==3.11.0
psycopg2-binary==2.8.4
celery[redis]==4.4.0
faker==4.0.1
